+++ estnltk/storage/postgres/queries/missing_layer_query.py
@@ -3,6 +3,7 @@ from typing import Set
 from psycopg2.sql import SQL, Identifier
 
 from estnltk.storage.postgres import layer_table_identifier
+from estnltk.storage.postgres import collection_table_identifier
 from estnltk.storage.postgres.queries.query import Query
 
 
@@ -18,11 +19,14 @@ class MissingLayerQuery(Query):
     def eval(self, collection: 'PgCollection'):
         collection_name = collection.name
         storage = collection.storage
+        collection_identifier = collection_table_identifier(collection.storage, collection.name)
         table = layer_table_identifier(storage, collection_name, self.missing_layer)
-        pat = SQL('{collection_name}."id" NOT IN (SELECT "text_id" FROM {table})').format(
-            collection_name = Identifier(collection_name),
+        # Note: the previous solution ( WHERE id NOT IN get_all_ids_from_missing_layer_table ) was terribly slow 
+        #       on large tables. Credits for the optimization hint go to: https://stackoverflow.com/a/16996103 
+        pat = SQL('NOT EXISTS (SELECT 1 FROM {table} WHERE {collection_identifier}."id" = {table}."text_id" LIMIT 1)').format(
+            collection_identifier = collection_identifier,
             table=table
-            )
+        )
         return pat
 
     @property