+++ estnltk/taggers/standard_taggers/flatten_tagger.py
@@ -22,6 +22,14 @@ class FlattenTagger(Tagger):
         self.attribute_mapping = attribute_mapping
         self.default_values = default_values
 
+    def _make_layer_template(self):
+        return Layer(name=self.output_layer,
+                     attributes=self.output_attributes,
+                     text_object=None,
+                     parent=None,
+                     enveloping=None,
+                     ambiguous=False)
+
     def _make_layer(self, text, layers, status):
         layer = flatten(input_layer=layers[self.input_layers[0]],
                         output_layer=self.output_layer,