+++ estnltk/taggers/syntax/stanza_tagger/ensemble_tagger.py
@@ -12,6 +12,8 @@ from estnltk.taggers.syntax.ud_validation.deprel_agreement_retagger import Depre
 from estnltk.taggers.syntax.ud_validation.ud_validation_retagger import UDValidationRetagger
 from estnltk.taggers.tagger import Tagger
 
+from estnltk.converters.serialisation_modules import syntax_v0
+
 RESOURCES = os.path.join(PACKAGE_PATH, 'taggers', 'syntax', 'stanza_tagger', 'stanza_resources')
 
 
@@ -104,6 +106,17 @@ class StanzaSyntaxEnsembleTagger(Tagger):
                 self.agreement_error_retagger = DeprelAgreementRetagger(output_layer=output_layer)
                 self.output_attributes += ('agreement_deprel',)
 
+    def _make_layer_template(self):
+        """Creates and returns a template of the layer."""
+        layer = Layer(name=self.output_layer,
+                      text_object=None,
+                      attributes=self.output_attributes,
+                      parent=self.input_layers[1],
+                      ambiguous=False )
+        if self.add_parent_and_children:
+            layer.serialisation_module = syntax_v0.__version__
+        return layer
+
     def _make_layer(self, text, layers, status):
         # Make an internal import to avoid explicit stanza dependency
         from stanza import Document
@@ -175,12 +188,8 @@ class StanzaSyntaxEnsembleTagger(Tagger):
 
         parent_layer = layers[self.input_layers[1]]
 
-        layer = Layer(name=self.output_layer,
-                      text_object=text,
-                      attributes=self.output_attributes,
-                      parent=self.input_layers[1],
-                      ambiguous=False,
-                      )
+        layer = self._make_layer_template()
+        layer.text_object=text
 
         extracted_words = [word for sentence in extracted_data for word in sentence]
         for line, span in zip(extracted_words, parent_layer):