+++ estnltk/taggers/text_segmentation/clause_segmenter.py
@@ -112,6 +112,14 @@ class ClauseSegmenter(Tagger):
             if self._java_process._process.poll() is None:
                 self.__exit__()
 
+    def _make_layer_template(self):
+        """Creates and returns a template of the layer."""
+        return Layer(name=self.output_layer,
+                      enveloping=self._input_words_layer,
+                      text_object=None,
+                      attributes=self.output_attributes,
+                      ambiguous=False)
+
     def _make_layer(self, text, layers, status: dict):
         """Tags clauses layer.
         
@@ -130,11 +138,8 @@ class ClauseSegmenter(Tagger):
            This can be used to store metadata on layer tagging.
 
         """
-        layer = Layer(name=self.output_layer,
-                      enveloping=self._input_words_layer,
-                      text_object=text,
-                      attributes=self.output_attributes,
-                      ambiguous=False)
+        layer = self._make_layer_template()
+        layer.text_object = text
 
         clause_spanlists = []
         # Iterate over sentences and words, tag clause boundaries