+++ estnltk/taggers/text_segmentation/paragraph_tokenizer.py
@@ -22,16 +22,22 @@ class ParagraphTokenizer(Tagger):
         self.paragraph_tokenizer = RegexpTokenizer(regex, gaps=True, discard_empty=True)
         self.regex = regex
 
+    def _make_layer_template(self):
+        """Creates and returns a template of the layer."""
+        return Layer(name=self.output_layer,
+                     text_object=None,
+                     enveloping=self._input_sentences_layer, 
+                     ambiguous=False)
+
     def _make_layer(self, text, layers, status: dict):
         """Tag paragraphs layer.
         
         Paragraph can only end at the end of a sentence.
 
         """
-        layer = Layer(name=self.output_layer,
-                      text_object=text,
-                      enveloping=self._input_sentences_layer, 
-                      ambiguous=False)
+        layer = self._make_layer_template()
+        layer.text_object = text
+        
         paragraph_ends = {end for _, end in self.paragraph_tokenizer.span_tokenize(text.text)}
         start = 0
         if layers[ self._input_sentences_layer ]: